requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openpyxl>=3.1.0
pydantic>=2.0.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.0
